Global Hashmap between order_id -> *Order, working O(1) order cancelation. removeOrderById(n), bid & asks are sorted accordingly (accending/decending) commiting from 10k feet

+++ src/book/order.zig
@@ -6,14 +6,14 @@ pub const Side = enum(u8) {
 };
 
 pub const Order = struct {
-    order_id: u64, 
-    price: f64, 
-    quantity: u32, 
-    side: Side, 
-    timestamp: u64,    
+    order_id: u64,
+    price: f32,
+    quantity: u32,
+    side: Side,
+    timestamp: u64,
 
-    pub fn init(order_id: u64, price: f64, quantity: u32, side: Side, timestamp: u64) Order {
-        return Order {
+    pub fn init(order_id: u64, price: f32, quantity: u32, side: Side, timestamp: u64) Order {
+        return Order{
             .order_id = order_id,
             .price = price,
             .quantity = quantity,
